pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kauakirchner <kauakirchner2@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.0.0"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
flasgger = "^0.9.7.1"
requests = "^2.31.0"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
google-auth-oauthlib = "^1.2.0"
google-auth = "^2.29.0"
flask-cors = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
